--- app/src/main/java/com/stochanskyi/librariesdemo/presentaiton/activityrecognition/ActivityRecognitionViewModel.kt
@@ -1,7 +1,0 @@
-package com.stochanskyi.librariesdemo.presentaiton.activityrecognition
-
-import androidx.lifecycle.ViewModel
-import javax.inject.Inject
-
-class ActivityRecognitionViewModel @Inject constructor() : ViewModel() {
-}
