+++ app/src/main/java/com/stochanskyi/librariesdemo/presentaiton/feature/activityrecognition/ActivityRecognitionViewModel.kt
@@ -1,0 +1,7 @@
+package com.stochanskyi.librariesdemo.presentaiton.feature.activityrecognition
+
+import androidx.lifecycle.ViewModel
+import javax.inject.Inject
+
+class ActivityRecognitionViewModel @Inject constructor() : ViewModel() {
+}
