+++ app/src/main/java/com/stochanskyi/librariesdemo/presentaiton/feature/activityrecognition/activityupdate/viewdata/ActivityUpdateEventViewData.kt
@@ -1,0 +1,9 @@
+package com.stochanskyi.librariesdemo.presentaiton.feature.activityrecognition.activityupdate.viewdata
+
+import androidx.annotation.StringRes
+
+data class ActivityUpdateEventViewData(
+    val id: Int,
+    @StringRes val activityNameRes: Int,
+    val activityConfidence: Int
+)
