+++ app/src/main/java/com/stochanskyi/librariesdemo/presentaiton/feature/imageloading/models/ImageLoaderViewData.kt
@@ -1,0 +1,8 @@
+package com.stochanskyi.librariesdemo.presentaiton.feature.imageloading.models
+
+import androidx.annotation.StringRes
+
+data class ImageLoaderViewData(
+    @StringRes val nameRes: Int,
+    val id: Int
+)
