+++ app/src/main/java/com/stochanskyi/librariesdemo/presentaiton/feature/main/MainActivity.kt
@@ -1,0 +1,17 @@
+package com.stochanskyi.librariesdemo.presentaiton.feature.main
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import com.stochanskyi.librariesdemo.databinding.ActivityMainBinding
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        installSplashScreen()
+        super.onCreate(savedInstanceState)
+
+        ActivityMainBinding.inflate(layoutInflater).let {
+            setContentView(it.root)
+        }
+    }
+}
