--- app/src/main/java/com/stochanskyi/librariesdemo/presentaiton/imageloading/models/ImageLoaderViewData.kt
@@ -1,8 +1,0 @@
-package com.stochanskyi.librariesdemo.presentaiton.imageloading.models
-
-import androidx.annotation.StringRes
-
-data class ImageLoaderViewData(
-    @StringRes val nameRes: Int,
-    val id: Int
-)
