+++ app/src/main/res/layout/fragment_activity_update.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:paddingTop="30dp"
-    tools:context=".presentaiton.main.MainActivity">
+    tools:context=".presentaiton.feature.main.MainActivity">
 
     <TextView
         style="@style/Text"
